# Conflicts:
#	src/main/java/com/cydeo/javahedgehogsproject/controller/SalesInvoiceController.java
#	src/main/java/com/cydeo/javahedgehogsproject/repository/InvoiceRepository.java
#	src/main/java/com/cydeo/javahedgehogsproject/service/InvoiceProductService.java
#	src/main/java/com/cydeo/javahedgehogsproject/service/InvoiceService.java
#	src/main/java/com/cydeo/javahedgehogsproject/service/implementation/InvoiceProductServiceImpl.java
#	src/main/java/com/cydeo/javahedgehogsproject/service/implementation/InvoiceServiceImpl.java

+++ src/main/java/com/cydeo/javahedgehogsproject/dto/InvoiceProductDto.java
@@ -14,7 +14,7 @@ public class InvoiceProductDto {
     private Long id;
     private Integer quantity;
     private BigDecimal price;
-    private Integer tax;
+    private BigDecimal tax;
     private BigDecimal total;
     private BigDecimal profitLoss;
     private Integer remainingQty;
